Links #45 moved stuff to svg
