+++ ng/projects/gongsvgspecific/src/lib/svg/svg.component.css
@@ -1,0 +1,12 @@
+.parent {
+    position: relative;
+    width: 3000;
+}
+
+.child {
+    position: absolute;
+    width: 3000px;
+    height: 8000px;
+    z-index: 100;
+    pointer-events: none;
+}
