+++ ng/projects/gongsvgspecific/src/lib/svg/svg.component.html
@@ -1,0 +1,14 @@
+<div class="parent">
+    <svg *ngIf="linkDrawing" class="child">
+        <g>
+            <line [attr.x1]="startX" [attr.y1]="startY" [attr.x2]="endX" [attr.y2]="endY" stroke="black" stroke-dasharray="5,5" [attr.stroke-width]="2">
+            </line>
+        </g>
+    </svg>
+</div>
+<div *ngFor="let layer of svg.Layers" class="parent">
+    <div *ngFor="let elt of svg.Layers" class="parent">
+        <lib-layer [GONG__StackPath]="GONG__StackPath" [Layer]="elt">
+        </lib-layer>
+    </div>
+</div>
